Changed redux setup to user new reducer and actions for user snippets
