+++ src/reducers/app.js
@@ -132,12 +132,6 @@ const app = (state = initialState, action) => {
         filters,
       };
     }
-    case actions.SAVE_USER_SNIPPETS: {
-      return {
-        ...state,
-        snippetMeta: action.payload,
-      }
-    }
     default: {
       return state;
     }
