+++ src/reducers/index.js
@@ -1,14 +1,16 @@
-import { combineReducers } from 'redux'
-import app from './app'
-import annotation from './annotation'
-import parser from './parser'
-import permissions from './permissions'
+import { combineReducers } from 'redux';
+import app from './app';
+import annotation from './annotation';
+import parser from './parser';
+import permissions from './permissions';
+import user from './user';
 
 const reducer = combineReducers({
   app,
   annotation,
   parser,
-  permissions
+  permissions,
+  user,
 })
 
 export default reducer
